Move error conditions near their conditions
This was an interesting exercise. The process was actually very
straightforward, but only because I had read the Ruby Style Guide
recently. I remembered the style guide to fail out of a method
as soon as possible. That meant that I should move the code in the
outer else statement to the top to fail immediately if the
condition was met. I saw that the condition needed to be changed
from expecting true to expecting false, so I changed it to unless
instead of if. I then did the same two steps for the inner if
statement.