Consolidate char and num generator methods
There was no need to call the char and num generators outside of
the name generator, so I moved all the code into one (still short)
method. The trouble right now is that the exception handling in
the initialize method is really more related to generating the
name. The next step will be to seperate that error handling out.