Use || operator instead of if statement
The if statement to check whether a generator had been passed in
was too wordy. Using || on the one line was much more consise.
Also, it made more sense this way because then @name_generator
will always hold the generator object that was used to generate
the robot name regardless of whether or not the generator was
passed in.